EES-60 - made tests more resilient to additional seed data on pages, by making test ids more explicit
