+++ src/explore-education-statistics-common/src/components/AccordionSection.tsx
@@ -1,3 +1,4 @@
+import {formatTestId} from "@admin/util/test-utils";
 import useMounted from '@common/hooks/useMounted';
 import findAllParents from '@common/lib/dom/findAllParents';
 import classNames from 'classnames';
