+++ tests/robot-tests/tests/admin/edit_release_page.robot
@@ -1,7 +1,7 @@
 *** Settings ***
 Resource    ../libs/library.robot
 
-Force Tags  Admin
+Force Tags  Admin  NotAgainstProd
 
 Suite Setup       user signs in
 Suite Teardown    user closes the browser
@@ -19,11 +19,12 @@ User selects theme and topic from dropdowns
 
 User clicks edit release
     [Tags]  HappyPath
-    user clicks element   css:#publications-2-heading
-    user waits until page contains element  xpath://dt[text()="Releases"]
-    user clicks element  css:#publications-2-content [data-testid="details--expand"]
-    user waits until page contains element  xpath://a[text()="Edit this release"]
-    user clicks link  Edit this release
+    user checks page contains accordion  Automated Test Publication for Edit Release
+    user opens accordion section  Automated Test Publication for Edit Release
+    user checks page contains details section  Academic Year, 2018 to 2019 (not Live)
+    user opens details section  Academic Year, 2018 to 2019 (not Live)
+    user waits until page contains element  css:[data-testid="Edit release link for Automated Test Publication for Edit Release, Academic Year, 2018 to 2019 (not Live)"]
+    user clicks element  css:[data-testid="Edit release link for Automated Test Publication for Edit Release, Academic Year, 2018 to 2019 (not Live)"]
 
 Validate release summary tab has correct details
     [Tags]  HappyPath
@@ -36,16 +37,16 @@ Validate release summary tab has correct details
     user waits until page contains element   xpath://dt[text()="Next release expected"]
     user waits until page contains element   xpath://dt[text()="Release type"]
 
-User clicks Edit release setup details, checks details and cancels
+User clicks Edit release summary, checks details and cancels
     [Tags]  HappyPath
-    user clicks link  Edit release setup details
+    user clicks link  Edit release summary
     user waits until page contains element   xpath://h2[text()="Edit release summary"]
     user checks element attribute value should be  css:#releaseSummaryForm-timePeriodCoverageStartYear   value   2018
     user checks radio option should be  releaseSummaryForm-releaseTypeId  Official Statistics
     user clicks link  Cancel update
     user waits until page contains element   xpath://h2[text()="Release summary"]
 
-User clicks Edit release setup details, changes details and updates
+User clicks Edit release summary, changes details and updates
     [Tags]  HappyPath  AltersData
-    user waits until page contains element  xpath://a[text()="Edit release setup details"]
-    user clicks link  Edit release setup details
+    user waits until page contains link  Edit release summary
+    user clicks link  Edit release summary
