+++ tests/robot-tests/tests/admin/login.robot
@@ -3,7 +3,7 @@ Resource    ../libs/library.robot
 
 Force Tags  Admin
 
-Suite Setup       user signs in
+Suite Setup       user opens the browser
 Suite Teardown    user closes the browser
 
 *** Test Cases ***
